Create api get and post for user
