+++ server/.env
@@ -1,2 +1,2 @@
-PORT = 5000
+PORT = 5001
 MONGO_DB = mongodb+srv://laonhi100:0966689312LUC@osm-shop.2wnnbil.mongodb.net/?retryWrites=true&w=majority
