+++ server/routes/coreRoutes/productRoutes.js
