--- server/routes/index.js
@@ -1,8 +1,0 @@
-const UserRoutes = require('./UserRouter')
-const routes = (app) => {
-    app.use('/api/user', (req, res) => {
-        res.send('User page')
-    })
-}
-
-module.exports = routes;
