--- server/services/UserService.js
@@ -1,13 +1,0 @@
-const createUser = () => {
-    return new Promise((resolve, reject) => {
-        try {
-            resolve({})
-        } catch (e) {
-            reject(e)
-        }
-    })
-}
-
-module.exports = {
-    createUser
-}
